RESOLVE: Issue51 tabs
