+++ src/ui/less/typography.less
@@ -33,7 +33,7 @@ a:hover {
   color: #C7692B;
 }
 
-p {
+p, ul, li {
   font-size: 1.02rem;
   line-height: 1.66rem;
   margin-bottom: 1.66rem;
